Fix native Win32 RECT dimension type mismatches

+++ Client/TeamTalkClassic/NativeParityFeatures.cpp
@@ -250,7 +250,7 @@ namespace
         int left = row.left;
         int right = groupRc.right - 8;
         int width = (std::max)(70, (right - left - gap) / 2);
-        int height = (std::max)(20, row.bottom - row.top + 6);
+        int height = (std::max)(20, static_cast<int>(row.bottom - row.top + 6));
 
         MoveWindow(official, left, row.top, width, height, TRUE);
         state->checkbox = CreateWindowExW(
@@ -422,7 +422,9 @@ namespace
             RECT rc = {};
             GetClientRect(list, &rc);
             int itemHeight = static_cast<int>(SendMessageW(list, LB_GETITEMHEIGHT, 0, 0));
-            int visible = itemHeight > 0 ? (std::max)(1, (rc.bottom - rc.top) / itemHeight) : 1;
+            int visible = itemHeight > 0
+                ? (std::max)(1, static_cast<int>(rc.bottom - rc.top) / itemHeight)
+                : 1;
             SendMessageW(list, LB_SETTOPINDEX, (std::max)(0, count - visible), 0);
         }
 
@@ -461,8 +463,8 @@ namespace
             return;
         RECT rc = ChildRect(hwnd, state->rich);
         MoveWindow(state->list, rc.left, rc.top,
-                   (std::max)(1, rc.right - rc.left),
-                   (std::max)(1, rc.bottom - rc.top), TRUE);
+                   (std::max)(1, static_cast<int>(rc.right - rc.left)),
+                   (std::max)(1, static_cast<int>(rc.bottom - rc.top)), TRUE);
     }
 
     void ApplyChatHistory(HWND hwnd)
@@ -476,8 +478,8 @@ namespace
             WS_EX_CLIENTEDGE, L"LISTBOX", L"Chat history",
             WS_CHILD | WS_VISIBLE | WS_TABSTOP | WS_VSCROLL |
             LBS_NOTIFY | LBS_NOINTEGRALHEIGHT,
-            rc.left, rc.top, (std::max)(1, rc.right - rc.left),
-            (std::max)(1, rc.bottom - rc.top), hwnd,
+            rc.left, rc.top, (std::max)(1, static_cast<int>(rc.right - rc.left)),
+            (std::max)(1, static_cast<int>(rc.bottom - rc.top)), hwnd,
             reinterpret_cast<HMENU>(static_cast<INT_PTR>(IDC_NATIVE_HISTORY_LIST)),
             AfxGetInstanceHandle(), nullptr);
         if (!state->list)
